bit value of index

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -10,7 +10,7 @@ int get_bit(unsigned long int n, unsigned int index)
 {
 	unsigned int i;
 
-	if (index > sizeof(size_t) * 8)
+	if (index > sizeof(unsigned int) * 8)
 		return (-1);
 	for (i = 0; i < index; i++)
 		n = n >> 1;
